Support pull_request_target
`pull_request_target` is granted a read/write repository token and runs
in the context of the base of the pull request. It allows to create
workflows that label and comment on pull requests based on the contents
of the event payload. The `get-changed-files` action is useful for some
of those actions and the code works for this case with almost no
modification.